Add backtrace and post hw HARU emulation and mapq computation for accurate mapping

+++ src/shared.h
@@ -5,9 +5,9 @@
 #include <stdint.h>
 
 #define TEST_SCALING
-typedef int16_t SIG_DTYPE;
-typedef uint16_t COST_DTYPE;
-#define COST_DTYPE_MAX UINT16_MAX
+typedef int32_t SIG_DTYPE;
+typedef uint32_t COST_DTYPE;
+#define COST_DTYPE_MAX UINT32_MAX
 #define SCALING 32.0
 
 #endif // SHARED_H
